+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## 0.11.1
 ### Fixed
 - Login not possible on instances in subfolder
+- Last editor not visible in _Entity Detail_ tab
 
 ## 0.11 - Kilcrea
 ### Added