+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## 0.12 -
+### Added
+- Two-Factor authentication
+
 ## 0.11 - Kilcrea
 ### Added
 - Added websockets for various data synchronization